+++ sdk/dotnet/Folder.cs
@@ -28,7 +28,7 @@ namespace Pulumi.VSphere
         /// for a reference on how to set values for custom attributes.
         /// </summary>
         [Output("customAttributes")]
-        public Output<ImmutableDictionary<string, object>?> CustomAttributes { get; private set; } = null!;
+        public Output<ImmutableDictionary<string, string>?> CustomAttributes { get; private set; } = null!;
 
         /// <summary>
         /// The ID of the datacenter the folder will be created in.
@@ -107,16 +107,16 @@ namespace Pulumi.VSphere
     public sealed class FolderArgs : Pulumi.ResourceArgs
     {
         [Input("customAttributes")]
-        private InputMap<object>? _customAttributes;
+        private InputMap<string>? _customAttributes;
 
         /// <summary>
         /// Map of custom attribute ids to attribute 
         /// value strings to set for folder. See [here][docs-setting-custom-attributes]
         /// for a reference on how to set values for custom attributes.
         /// </summary>
-        public InputMap<object> CustomAttributes
+        public InputMap<string> CustomAttributes
         {
-            get => _customAttributes ?? (_customAttributes = new InputMap<object>());
+            get => _customAttributes ?? (_customAttributes = new InputMap<string>());
             set => _customAttributes = value;
         }
 
@@ -164,16 +164,16 @@ namespace Pulumi.VSphere
     public sealed class FolderState : Pulumi.ResourceArgs
     {
         [Input("customAttributes")]
-        private InputMap<object>? _customAttributes;
+        private InputMap<string>? _customAttributes;
 
         /// <summary>
         /// Map of custom attribute ids to attribute 
         /// value strings to set for folder. See [here][docs-setting-custom-attributes]
         /// for a reference on how to set values for custom attributes.
         /// </summary>
-        public InputMap<object> CustomAttributes
+        public InputMap<string> CustomAttributes
         {
-            get => _customAttributes ?? (_customAttributes = new InputMap<object>());
+            get => _customAttributes ?? (_customAttributes = new InputMap<string>());
             set => _customAttributes = value;
         }